% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgmh.R
\name{cgmh}
\alias{cgmh}
\title{Wrapper function for C code.}
\usage{
cgmh(A, Y, N)
}
\arguments{
\item{A}{Vector of returned acceptance values.}

\item{Y}{Vector of proposed moves.}

\item{N}{Number of parallel samples.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Wrapper function for C code.
}

