% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GPMH-internal.R
\docType{package}
\name{GPMH}
\alias{GPMH}
\alias{GPMH-package}
\title{GPMH: A package for computating metropolis hastings in parallel.}
\description{
The GPMH package implements the algorithm for a general parallel metropolis
hastings by Calderhead using various technologies for comparison.
}

