% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MH.R
\name{MH}
\alias{MH}
\title{Metropolis-Hastings algorithm}
\usage{
MH(target, kernel, dkernel, init.state, n)
}
\arguments{
\item{target}{The target distribution that the algorithm aims at sampling from.}

\item{kernel}{Input function to draw samples from the proposal distribution.}

\item{dkernel}{Input function to evaluate the proposal distribution pointwise.}

\item{init.state}{Sets the starting value for the first point of the Markov Chain.}

\item{n}{Number of total iterations of the algorithm.}
}
\value{
Returns a vector of \code{n} samples from the target distribution.
}
\description{
Runs the standard Metropolis-Hastings algorithm for sampling
 from a given target distribution.
}
\details{

}
\examples{
######MH######
xMH<-MH(target=function(x) {0.2*dnorm(x,12,1)+0.3*dnorm(x,8,1)+0.4*dnorm(x,4,1)+0.6*dnorm(x,15,1)+0.3*dnorm(x,20,1)},
         kernel=function(x) {rnorm(n=1, x)},
         dkernel=function(x,y) {dnorm(y, x)},
         init.state=5,
         n=100000)

plot(xMH$x)
summary(xMH$x)
}

