% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rcGPMH.R
\name{rcGPMH}
\alias{rcGPMH}
\title{Generalised Metropolis-Hastings algorithm (R/C version)}
\usage{
rcGPMH(target, kernel, dkernel, init.state, n, N = 8)
}
\arguments{
\item{target}{The target distribution that the algorithm aims at sampling from.}

\item{kernel}{Input function to draw samples from the proposal distribution.}

\item{dkernel}{Input function to evaluate the proposal distribution pointwise.}

\item{init.state}{Sets the starting value for the first point of the Markov Chain.}

\item{n}{Number of total iterations of the algorithm.}

\item{N}{Number of samples to draw from the proposal distribution at each step.}
}
\value{
Returns a vector of \code{n*(N+1)} samples from the target distribution
}
\description{
Runs the Generalised Metropolis-Hastings algorithm for sampling
  from a given target distribution. At each iteration, the algorithm draws N
  new points from the proposal distribution given the current state, then it
  calls the C function "Cgmh.c" to compute the likelihoods of all N+1 points
  (the N new points plus the current state). Then, these N+1 points are
  sampled with replacement according to their likelihoods and the new sample
  of N+1 points is obtained. Finally, a state from these N+1 points is
  randomly chosen to generate the N new points in the successive iteration.
}
\examples{
######rcGMH######
###R/C
xrcGMH<-rcGPMH(target=function(x) {0.2*dnorm(x,12,1)+0.3*dnorm(x,8,1)+0.4*dnorm(x,4,1)+0.6*dnorm(x,15,1)+0.3*dnorm(x,20,1)},
              kernel=function(x) {rnorm(n=1, x)},
              dkernel=function(x,y) {dnorm(y, x)},
              init.state=5,
              n=100000,
              N=8)

plot(xrcGMH$x)
summary(xrcGMH$x)

}

